style: 변수이름 수정
